% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{appendWithName}
\alias{appendWithName}
\title{Append Elements to a List by Name}
\usage{
appendWithName(lst_, ...)
}
\arguments{
\item{lst_}{A list to which new elements will be appended.}

\item{...}{Named arguments representing the elements to be appended to the list.}
}
\value{
A list with the new elements appended.
}
\description{
This function appends new elements to an existing list, using the provided names for the new elements.
}
\examples{
my_list = list(a = 1, b = 2)
my_list = appendWithName(my_list, c = 3, d = 4)
print(my_list)

}
