% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asColoredChr}
\alias{asColoredChr}
\title{Convert Character Vector to Colored Character Class}
\usage{
asColoredChr(vec_)
}
\arguments{
\item{vec_}{A character vector to be converted.}
}
\value{
A character vector with the class attribute `coloredChr`.
}
\description{
This function converts a character vector to a special class `coloredChr`, which is used to store and print color data.
}
\examples{
vec = c("red", "green", "blue")
colored_vec = asColoredChr(vec)
class(colored_vec)  # "coloredChr"

}
