% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{biasedMap}
\alias{biasedMap}
\title{Apply a Function to Specific Elements of a List}
\usage{
biasedMap(lst_, fun_, luckyOnes_ = 1, reverse_ = F, ...)
}
\arguments{
\item{lst_}{A list whose elements the function will be applied to.}

\item{fun_}{A function to apply to the elements.}

\item{luckyOnes_}{An integer vector indicating the indices of the lucky elements to apply the function to.}

\item{reverse_}{A logical value. If `TRUE`, the function is applied to the unlucky elements instead. Defaults to `FALSE`.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A list with the function applied to the specified elements.
}
\description{
This function applies a specified function to either a subset of elements (lucky ones) or the rest (unlucky ones) in a list.
}
\examples{
my_list = list(1, 2, 3, 4)
biasedMap(my_list, \(x) x^2, luckyOnes_ = c(2, 3))

}
