% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFitnessScores.R
\name{calculateFitnessScores}
\alias{calculateFitnessScores}
\title{Calculate Fitness Scores}
\usage{
calculateFitnessScores(DMS_data, nonselect_cutoff = 200)
}
\arguments{
\item{DMS_data}{Data frame FOR A SINGLE GENE of the form output by \code{readDMS}}

\item{nonselect_cutoff}{Filter out low confidence variants with less than this many reads per million in nonselective conditions. (Default is 200 reads per million)}
}
\value{
Input data frame plus the following columns: select (the average of select1 and select2), nonselect (the average of nonselect1 and nonselect2), fold_change, fitness_score
}
\description{
Calculate fitness scores of variants from raw DMS data.
}
\details{
A more detailed description of what the function does.
}
\examples{
# this example only works on my computer
setwd("/Users/Rachel/desktop/Roth Lab/R_DMS")
path <- './data/rawData_UBE2I_solid.txt'
gene_name <- 'UBE2I'
raw_data <- readDMS(path, gene_name)
head(calculateFitnessScores(raw_data))
}
\seealso{
\code{\link{readDMS}} The output of this function should be used as input for calculateFitnessScores.
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
