% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{clean_phe}
\alias{clean_phe}
\title{Clean the phenotype data from TCGA}
\usage{
clean_phe(phe, min_survival_time = 30)
}
\arguments{
\item{phe}{data.frame, the raw clinical information including vital_status, days_to_last_follow_up and days_to_death columns.}

\item{min_survival_time}{integer, the threshold to delete individual with survival time < 30 days}
}
\value{
data.frame, the clean clinical information including survivalStatus and survivalTime columns.
}
\description{
Given a phenotype data frame, delete the individual with NA and calculate the survival time and status.
}
\details{
nothing
}
\examples{
data(phe);phe = clean_phe(phe)
}
