% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{cleanphe}
\alias{cleanphe}
\title{Clean the phenotype data from TCGA}
\usage{
cleanphe(phe)
}
\arguments{
\item{phe}{data.frame, the raw clinical information including vital_status, days_to_last_follow_up and days_to_death columns.}
}
\value{
data.frame, the clean clinical information including survivalStatus and survivalTime columns.
}
\description{
Given a phenotype data frame, delete the individual with NA and calculate the survival time and status.
}
\details{
nothing
}
\examples{
data(phe);phe = cleanphe(phe)
}
