% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{cluster2major}
\alias{cluster2major}
\title{Map clusters to cell types}
\usage{
cluster2major(..., nCluster_, remain2_ = NULL)
}
\arguments{
\item{...}{Expressions that map cluster numbers to cell types.}

\item{nCluster_}{Integer. The total number of clusters.}

\item{remain2_}{Integer. The target group for unassigned clusters. Defaults to the first empty group if NULL.}
}
\value{
A named vector where keys are cluster numbers, and values are corresponding cell type labels.
}
\description{
Assigns cluster numbers to cell type labels. Unassigned clusters are grouped into the remaining group.
}
\examples{
cluster2cell(cluster_A = c(0, 1), cluster_B = c(2, 3), nCluster_ = 5)
}
