% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{combinePheExpr}
\alias{combinePheExpr}
\title{Clean the pheontype and expression data from TCGA}
\usage{
combinePheExpr(phe, expr)
}
\arguments{
\item{phe}{A data.frame, the raw clinical information including vital_status, days_to_last_follow_up and days_to_death columns.}

\item{expr}{A data.frame, the raw expression matrix with colname of molecular id and individual ids.}
}
\value{
A data.frame include clinical information and expression matrix
}
\description{
Clean the pheontype and expression data from TCGA
}
\details{
nothing
}
\examples{
\dontrun{
library(bioTools)
cleanData = clean(phe,expr)
}
}
