% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{count2tpm}
\alias{count2tpm}
\title{Convert Raw Counts to TPM (Transcripts Per Million)}
\usage{
count2tpm(mat_, org_ = "hsa", keytype_ = "symbol")
}
\arguments{
\item{mat_}{A numeric matrix of raw gene counts, where rows are genes and columns are samples.}

\item{org_}{A character string indicating the organism, either 'hsa' for human or 'mmu' for mouse (default is 'hsa').}

\item{keytype_}{A character string specifying the keytype to map gene IDs. Options are 'ensembl', 'entrez', or 'symbol' (default is 'symbol').}
}
\value{
A numeric matrix of TPM values (genes in rows, samples in columns).
}
\description{
This function converts raw gene count matrix into TPM values, using gene length information and effective length from a reference dataset.
The reference dataset is organism-specific (human or mouse), and the function uses the specified keytype (e.g., Ensembl, Entrez, or Symbol) to map gene IDs.
}
\examples{
mat_count = read.csv(fileWhenTest('bulk/count.csv', 'bioTool'), row.names = 'symbol')
mat_tpm = count2tpm(mat_count)

}
