% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{createGSEAinput}
\alias{createGSEAinput}
\title{create gct file and cls files}
\usage{
createGSEAinput(exprSet, groupList, prefix = "GSEA", destDir = ".")
}
\arguments{
\item{exprSet}{a expression matrix, which columns are sample,rows are HUGO gene symbols, or probeset ID .}

\item{groupList}{a vector,as long as the col number for the expression matrix,which describe the group for the samples in exprSet}

\item{prefix}{The prefix for gct file and cls files.}

\item{destdir}{where to store the files just download.}
}
\value{
A list of writed 2 files which are the input for GSEA (gct and cls format)
* `gctFile` - The saved path of GCT format file
* `clsFile` - The saved path of CLS format file
}
\description{
create gct file and cls file for GSEA according the expression matrix and group information
}
\examples{

createGSEAinput(exprSet = matrix(rnorm(1000), ncol = 10, nrow = 100), groupList = c(1, 1, 1, 1, 1, 2, 2, 2, 2, 2), 'GSE1009')
}
