% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doCal}
\alias{doCal}
\title{Execute a Function in the Context of an Object}
\usage{
doCal(x_, fun_, ...)
}
\arguments{
\item{x_}{An object that provides the context for evaluating the parameters.}

\item{fun_}{A function to be executed.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
The result of the function execution.
}
\description{
This function evaluates a specified function with additional parameters evaluated in the context of an object.
}
\examples{
my_env = new.env()
my_env$a = 2
my_env$b = 3
my_env |> doCal(sum, a, b)

}
