% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOParser.R
\name{downGEODatasets}
\alias{downGEODatasets}
\title{Download GEO datasets from Gene Expression Omnibus databse}
\usage{
downGEODatasets(geoIDs, log2Trans = FALSE, verbose = TRUE)
}
\arguments{
\item{geoIDs}{A vector of GSE accession number.}

\item{log2Trans}{log2(X + 1) transfer for the expression data, default: FALSE.}

\item{verbose}{Show running infos or not, default: TRUE.}
}
\value{
A list of downloaded gene expression profiles of each GSE accession number.
}
\description{
Download a list of GEO datasets from Gene Expression Omnibus databse.
}
\examples{

exprs.lst <- downGEODatasets(geoIDs = c('GSE19830'), log2Trans = FALSE)
}
