% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcgaParser.R
\name{downloadTcgaData}
\alias{downloadTcgaData}
\title{Fetch genomic datasets based on the given cancer types.}
\usage{
downloadTcgaData(
  cancerTypes = getGDCprojects()$project_id[grepl("TCGA", getGDCprojects()$project_id)],
  ...
)
}
\arguments{
\item{cancerTypes}{Specify the cancer types for downloading the genimic datasets from the TCGA database.}

\item{...}{Other parameters can be obtained through GDCdownload and GDCquery functions.}
}
\value{
A list of download datasets for the given cancer types.
* `data` - Fetched data matrix.
* `clinical` - Clinical information corressponding to the cancer type.
}
\description{
Fetch the genomic datasets for the given cancer type from TCGA database using the package 'TCGAbiolinks'. All cancer type names can be obtained from https://bioconductor.org/packages/release/bioc/html/TCGAbiolinks.html.
}
\examples{

downloadTcgaData(cancerTypes = c('TCGA-GBM'), data.category = 'Gene expression', data.type = 'Gene expression quantification', platform = 'Illumina HiSeq', file.type  = 'normalized_results', experimental.strategy = 'RNA-Seq', legacy = TRUE, directory = '.')
}
