% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_gsea.R
\name{enrich_gsea}
\alias{enrich_gsea}
\title{enrich_gsea}
\usage{
enrich_gsea(
  genelist,
  species = "hs",
  keytype = "symbol",
  db = "GO",
  go_ontology = "BP",
  pvalue_cutoff = 0.1,
  padjust_method = "BH",
  min_size = 10,
  max_size = 400,
  ...
)
}
\arguments{
\item{genelist}{character vector, must be gene symbols}

\item{species}{character, must be one of hs, mm}

\item{keytype}{character, indicating genes type, one of entrezid, symbol}

\item{db}{character, which internal db to enrich, one of GO, KEGG}

\item{go_ontology}{which GO ontology to use, one or more of BP, MF, CC,
only works if db contains GO}

\item{pvalue_cutoff}{pvalueCutoff}

\item{padjust_method}{pvalue_cutoff method, default is "BH"}

\item{min_size}{min set size}

\item{max_size}{max set size}

\item{...}{additional parameters pass to clusterProfiler::enricher}
}
\value{
enrichResult from enricher
}
\description{
GSEA enrichment using annoHub db
}
