% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichActSup}
\alias{enrichActSup}
\title{Enrichment Analysis for Activated and Suppressed Genes}
\usage{
enrichActSup(df_, toGo_ = "GO", cutoff_ = 0.05, organism_ = "hsa")
}
\arguments{
\item{df_}{A data frame containing gene data with at least `logFC` and `symbol` columns.}

\item{toGo_}{A string specifying the type of analysis: `'GO'` or `'KEGG'`. Defaults to `'GO'`.}

\item{cutoff_}{A numeric value specifying the FDR cutoff. Defaults to `0.05`.}

\item{organism_}{A string specifying the organism. `'hsa'` for human and `'mmu'` for mouse. Defaults to `'hsa'`.}
}
\value{
A list containing two enrichment analysis results: `act` (activated) and `sup` (suppressed).
}
\description{
This function performs separate enrichment analyses for activated and suppressed genes in a dataset.
}
\examples{

df = data.frame(symbol = c("TP53", "BRCA1", "MYC"), logFC = c(1.2, -0.5, 0.8))
enrich_actsup = enrichActSup(df, toGo_ = "GO", cutoff_ = 0.05)

}
