% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichment_of}
\alias{enrichment_of}
\title{Perform Easy Enrichment Analysis}
\usage{
enrichment_of(genes_, toGo_ = "GO", cutoff_ = 0.05, organism_ = "hsa")
}
\arguments{
\item{genes_}{A vector of gene symbols to analyze.}

\item{toGo_}{A string specifying the type of analysis: `'GO'` or `'KEGG'`. Defaults to `'GO'`.}

\item{cutoff_}{A numeric value specifying the FDR cutoff. Defaults to `0.05`.}

\item{organism_}{A string specifying the organism. `'hsa'` for human and `'mmu'` for mouse. Defaults to `'hsa'`.}
}
\value{
An enrichment analysis result object.
}
\description{
This function performs GO or KEGG enrichment analysis for a given set of genes.
}
\examples{
enrichment = enrichment_of(c("TP53", "BRCA1"))

}
