% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{env_meta_regr}
\alias{env_meta_regr}
\title{Environment for Regression Meta Information}
\format{
An object of class \code{environment} of length 2.
}
\usage{
env_meta_regr
}
\value{
An environment containing `meta_regr_builtin` and `meta_regr_custom`.
}
\description{
This object stores the environment for regression metric metadata, initialized by
`init_meta_regr()`. It contains both the built-in and custom regression metrics.
}
\keyword{datasets}
