% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_template.R
\name{example_function}
\alias{example_function}
\title{Function Template}
\usage{
example_function(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
The sum of x and y
}
\description{
Description of the function
}
\details{
A more detailed description of what the function does.
}
\examples{
# add 2 and 3
example_function(2,3)
# [1] 5


}
\seealso{
\code{\link{another_function}} Info
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
