% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_color}
\alias{get_color}
\title{Get a specified number of colors from a predefined dictionary of colors}
\usage{
get_color(n_, which_ = NULL)
}
\arguments{
\item{n_}{A numeric value specifying the number of colors to retrieve. Must be between 1 and 17.}

\item{which_}{A character string specifying which color to retrieve. Defaults to NULL, in which case all colors are retrieved.}
}
\value{
A colored vector of the requested number of colors.
}
\description{
This function retrieves `n_` colors from the color dictionary. If `n_` is greater than 3,
it retrieves from the `dict_color_n` dictionary; otherwise, it selects from the respective predefined color sets.
}
\examples{
get_color(3, 'morandi')
get_color(5, 'morandi')

}
