% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_color_2}
\alias{get_color_2}
\title{Get Color from dict_color_2, dict_color_3, and dict_color_n}
\usage{
get_color_2(which_ = "ddlc")
}
\arguments{
\item{which_}{A character string specifying the key to retrieve from the color dictionary. Defaults to 'ddlc'.}
}
\value{
A color corresponding to the specified key in the dictionary.
}
\description{
This function retrieves the color from the dictionary based on the specified `which_` parameter.
The default is 'ddlc'.
}
\keyword{internal}
