% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_color_3}
\alias{get_color_3}
\title{Get Color from dict_color_3 and dict_color_n}
\usage{
get_color_3(which_ = "htmap")
}
\arguments{
\item{which_}{A character string specifying the key to retrieve from the color dictionary. Defaults to 'htmap'.}
}
\value{
A color corresponding to the specified key in the dictionary.
}
\description{
This function retrieves the color from the dictionary based on the specified `which_` parameter.
The default is 'htmap'.
}
\keyword{internal}
