% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_color_n}
\alias{get_color_n}
\title{Get Color from dict_color_n}
\usage{
get_color_n(which_ = "macaron")
}
\arguments{
\item{which_}{A character string specifying the key to retrieve from the color dictionary. Defaults to 'macaron'.}
}
\value{
A color corresponding to the specified key in the dictionary.
}
\description{
This function retrieves the color from the dictionary based on the specified `which_` parameter.
The default is 'macaron'.
}
\keyword{internal}
