% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneset_go.R
\name{get_geneset_go}
\alias{get_geneset_go}
\title{get GO annotation}
\usage{
get_geneset_go(
  species = "hs",
  ontology = c("BP", "MF", "CC"),
  min_size = 5,
  max_size = 500,
  evidence_exclude = c("NAS", "NR", "ND", "IEA")
)
}
\arguments{
\item{species}{character, must be one of hs, mm}

\item{ontology}{character, GO ontology to keep, one or more of "BP","MF","CC"}

\item{min_size}{integer, minimal set size}

\item{max_size}{integer, maxium set size}

\item{evidence_exclude}{character, excluded GO evidence,
refer to http://geneontology.org/docs/guide-go-evidence-codes/}
}
\value{
tibble
}
\description{
get GO annotation from bioconductor org.Xx.eg.db
}
