% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneset_kegg.R
\name{get_geneset_kegg}
\alias{get_geneset_kegg}
\title{get KEGG annotation}
\usage{
get_geneset_kegg(species = "hs")
}
\arguments{
\item{species}{a character, must be one of hs, mm}
}
\value{
tibble
}
\description{
get KEGG annotation from KEGG API
}
