% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{getGO}
\alias{getGO}
\title{Retrieve GO Data}
\usage{
getGO(organism_ = "hsa", keyType_ = "SYMBOL")
}
\arguments{
\item{organism_}{A string specifying the organism. `'hsa'` for human and `'mmu'` for mouse. Defaults to `'hsa'`.}

\item{keyType_}{A string specifying the key type for gene identifiers. Defaults to `'SYMBOL'`.}
}
\value{
A list of GO data, including mappings of terms to genes and term names.
}
\description{
This function retrieves Gene Ontology (GO) data for a specified organism.
}
\examples{
GO = getGO()

}
