% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{gg_pca}
\alias{gg_pca}
\title{plot PCA figure}
\usage{
gg_pca(df, PC1 = "PC1", PC2 = "PC2", group = "group", explained_var = NULL)
}
\arguments{
\item{df}{A data.frame, the expression matrix in which column is multiple variables and row is individual.}

\item{PC1}{A character, the first eigenvalue}

\item{PC2}{A character, the second eigenvalue}

\item{group}{A character, one colname of phe variable.}

\item{explained_var}{A numeric vector.}
}
\value{
A ggplot object
}
\description{
plot PCA figure
}
\examples{
\dontrun{
library(bladderbatch)
data(bladderdata)
dat <- bladderEset[1:1000,]
edata <- exprs(dat)
pheno <- pData(dat)
pca_promp <- pca(t(edata))
explained_var <- pca_promp$sdev^2/sum(pca_promp$sdev^2)
df <- data.frame(pca_promp$x,pheno)
gg_pca(df,group="cancer",explained_var=explained_var)
}
}
