% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gzDir}
\alias{gzDir}
\title{Extract and Uncompress Gzipped Files in a Directory}
\usage{
gzDir(dir_)
}
\arguments{
\item{dir_}{A string specifying the directory containing `.gz` files.}
}
\value{
Invisible `NULL`.
}
\description{
This function uncompresses all `.gz` files in a specified directory.
}
\details{
- Removes the original `.gz` files after extraction.
- If no `.gz` files are found, prints "NULL" and does nothing.
}
\examples{
gzDir("path/to/directory")

}
