% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{hyperGtestTerms}
\alias{hyperGtestTerms}
\title{Hypergeometric Tests for GO/KEGG test}
\usage{
hyperGtestTerms(diffExprGenes, geneID2Path, path2GeneID, universeGeneIds)
}
\arguments{
\item{diffExprGenes}{a vector which contain the significantly DEG list.}

\item{geneID2Path}{a list which one entrez gene id to multiple pathway id.}

\item{path2GeneID}{a list which one pathway id to multiple entrezgene id.}

\item{universeGeneIds}{a vector which contain the backgroud gene list, probably 20,000 genes.}
}
\value{
a data.frame, each row is a Hypergeometric Tests result for each pathway.
}
\description{
Given GeneID2Path,Path2GeneID, diff_gene,universeGeneIds, this function will compute Hypergeometric Tests for each path (GO/KEGG),
It can't hold on the structure of the GO graph, just a simple path.
}
\examples{


}
