% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{init_metric_regr}
\alias{init_metric_regr}
\title{Environment for Regression Meta Information}
\usage{
init_metric_regr()
}
\description{
This object stores the environment for regression metric metadata, initialized by
`init_meta_regr()`. It contains both the built-in and custom regression metrics.
}
\keyword{internal}
