% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadSafely}
\alias{loadSafely}
\title{Safely Load Data from an RData File}
\usage{
loadSafely(path2RData_)
}
\arguments{
\item{path2RData_}{A string specifying the path to the RData file.}
}
\value{
A named list containing all objects from the RData file.
}
\description{
This function loads objects from an RData file into a list, avoiding pollution of the global environment.
}
\examples{
# Save and load an example RData file
my_data = list(a = 1, b = 2)
save(my_data, file = "example.RData")
loaded_data = loadSafely("example.RData")
print(loaded_data)

}
