% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{logNormalize}
\alias{logNormalize}
\title{Log-Normalize Gene Expression Data}
\usage{
logNormalize(mat_, log_ = log2, sf_ = 1)
}
\arguments{
\item{mat_}{A numeric matrix of gene expression values (genes in rows, samples in columns).}

\item{log_}{A function for the logarithmic transformation (default is `log2`).}

\item{sf_}{A scaling factor (default is 1), though it is not used in the current implementation.}
}
\value{
A numeric matrix of log-transformed gene expression values.
}
\description{
This function performs log-normalization on a gene expression matrix, adding a constant value (default is 1) to avoid taking the logarithm of zero.
The normalization is done element-wise, applying the specified logarithm function (default is `log2`).
}
\examples{
mat_count = read.csv(fileWhenTest('bulk/count.csv', 'bioTool'), row.names = 'symbol')
mat_tpm = count2tpm(mat_count)
mat_tpm = logNormalize(mat_tpm)

}
