% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsMethods.R, R/utils.R
\name{logTransform}
\alias{logTransform}
\title{Check numeric data needs log2-transformed or not based on the quantile method.}
\usage{
logTransform(X)

logTransform(X)
}
\arguments{
\item{X}{A data.frame or a vector of numeric data.}
}
\value{
Bool value. TRUE indicates the checked data is log-transformed, FALSE is not.

Bool value. TRUE indicates the checked data is log-transformed, FALSE is not.
}
\description{
Check numeric data needs log2-transformed or not based on the quantile method.

Check numeric data is log2-transformed or not based on the quantile method.
}
\examples{

logTransform(runif(n = 100, min = 1, max = 1000))

logTransform(runif(n = 100, min = 1, max = 1000))
}
