% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{lst2gmt}
\alias{lst2gmt}
\title{Convert a List to GMT Format}
\usage{
lst2gmt(lst_)
}
\arguments{
\item{lst_}{A named list where names are terms and values are vectors of associated genes.}
}
\value{
A tibble with two columns: `term` and `gene`.
}
\description{
This function converts a list of terms and associated genes into a GMT-like format.
}
\examples{
lst = list(Term1 = c("GeneA", "GeneB"), Term2 = c("GeneC"))
gmt = lst2gmt(lst)

}
