% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_TFdb.R
\name{make_TFdb}
\alias{make_TFdb}
\title{make TF to target dataset from online data}
\usage{
make_TFdb(category = c("trrust", "dorothea"), dorothea_confidence = c("A"))
}
\arguments{
\item{category}{must be c("trrust", "dorothea"), for now}

\item{dorothea_confidence}{dorothea filter level, "A" means validated at assay level}
}
\value{
a list containing two tibble: tf to target (hs and mm)
}
\description{

}
\examples{


}
