% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{mergeSeurat}
\alias{mergeSeurat}
\title{Merge Seurat objects}
\usage{
mergeSeurat(..., lst_ = NULL)
}
\arguments{
\item{...}{One or more Seurat objects to merge.}

\item{lst_}{List. A list of Seurat objects to merge. Defaults to NULL, which uses the objects passed through `...`.}
}
\value{
A merged Seurat object.
}
\description{
Combines multiple Seurat objects into a single Seurat object.
}
\examples{
mergeSeurat(obj1, obj2, obj3)

}
