% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\docType{data}
\name{metric_regr}
\alias{metric_regr}
\title{Environment for Regression Metric Functions}
\format{
An object of class \code{environment} of length 22.
}
\usage{
metric_regr
}
\description{
This object stores the environment for regression metric functions, initialized by
`init_metric_regr()`. It contains the functions to compute scores for built-in regression
metrics from `mlr3`.
}
\keyword{datasets}
