% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metrics_regr}
\alias{metrics_regr}
\title{Metrics Regression Function}
\usage{
metrics_regr()
}
\value{
A tibble containing the metadata of regression metrics.
}
\description{
A function to retrieve the list of regression metrics stored in the environment.
}
