\name{moco}
\alias{moco}
\docType{data}
\title{
Moco Cotton Data
}

\description{
Data set of...
}

\usage{data("moco")}
\format{
  A data frame with 206 observations (sampling points) on the following 20 variables (coordinates and markers).
  \describe{
    \item{\code{Lon}}{a numeric vector containing values of longitude}
    \item{\code{Lat}}{a numeric vector containing values of latitude}
    \item{\code{BNL1434.1}}{a numeric vector (marker)}
    \item{\code{BNL1434.2}}{a numeric vector}
    \item{\code{BNL840.1}}{a numeric vector}
    \item{\code{BNL840.2}}{a numeric vector}
    \item{\code{BNL2496.1}}{a numeric vector}
    \item{\code{BNL2496.2}}{a numeric vector}
    \item{\code{BNL1421.1}}{a numeric vector}
    \item{\code{BNL1421.2}}{a numeric vector}
    \item{\code{BNL1551.1}}{a numeric vector}
    \item{\code{BNL1551.2}}{a numeric vector}
    \item{\code{CIR249.1}}{a numeric vector}
    \item{\code{CIR249.2}}{a numeric vector}
    \item{\code{BNL3103.1}}{a numeric vector}
    \item{\code{BNL3103.2}}{a numeric vector}
    \item{\code{CIR311.1}}{a numeric vector}
    \item{\code{CIR311.2}}{a numeric vector}
    \item{\code{CIR246.1}}{a numeric vector}
    \item{\code{CIR246.2}}{a numeric vector}
  }
}

\source{
...
}

\references{
...
}

\examples{
data(moco)
str(moco)
}
\keyword{datasets}
