% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutateSeq.R
\name{mutateSeq}
\alias{mutateSeq}
\title{MutateSeq}
\usage{
mutateSeq(wt_sequence, start, stop, mut, wt = NULL)
}
\arguments{
\item{wt_sequence}{Character string representing the wild type DNA or RNA sequence to be mutated}

\item{start}{Integer nucleotide position that the mutation starts}

\item{stop}{Integer nucleotide position that the mutation ends (last mutant nucleotide)}

\item{mut}{Character vector of mutant sequence that the indicated positions should be changed to.}

\item{wt}{Optional character string, the part of the sequence to be replaced by the mutant sequence. Included only for checking purposes to make sure the position to mutate is correct.}
}
\value{
Character string of the mutant sequence.
}
\description{
Mutate a DNA or RNA sequence
}
\details{
Error if the wild type sequence at the postions provided does not match the wt seqment that you are trying to change.
Error if the mutant sequence provided does not match the length of the start and stop positions.
}
\examples{

mutateSeq("AAACATGGG", start = 4, stop = 6, wt = 'CAT', mut = 'GGC')
# [1] "AAAGGCGGG"


}
\seealso{
\code{\link{another_function}} Info
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
