% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutSeqs.R
\name{mutSeqs}
\alias{mutSeqs}
\title{Mutate Sequences}
\usage{
mutSeqs(sequences, start, stop, mut, wt = NULL)
}
\arguments{
\item{sequences}{Character vector representing the wild type DNA or RNA sequences to be mutated. Each sequences in a separate string.}

\item{start}{Integer vector, ith elements represents the nucleotide positions that the ith mutation starts.}

\item{stop}{Integer vector, ith element represents the nucleotide position that the ith mutation ends (last mutant nucleotide)}

\item{mut}{Character vector of mutant sequences that the indicated positions should be changed to.}

\item{wt}{Optional character vector, the part of the sequences to be replaced by the mutant sequences. In the same order as wt_sequences. Included only for checking purposes to make sure the position to mutate is correct.}
}
\value{
Character vector of the mutant sequences.
}
\description{
Mutate a vector of DNA or RNA sequences; vectorized.
}
\details{
Vectorized function: The ith sequences in the sequences vector will be mutated at the ith position in the start and stop vectors, and changed to the ith mutation provided in the mut vector.
Error if the wild type sequences at the postions provided does not match the wt seqment that you are trying to change.
Error if the mutant sequences provided does not match the length of the start and stop positions.
}
\examples{

mutSeqs(sequences = c('AAC', "GCC"), start = c(1,2), stop = c(2, 3), wt = c("AA", "CC"), mut = c('GG', 'AA'))

}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
