% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{performs a principal components analysis}
\usage{
pca(mat, ntop = 500, scale = FALSE)
}
\arguments{
\item{mat}{A data.frame, the expression matrix in which column is multiple variables and row is individual.}

\item{ntop}{A numeric, the top number of gene to performs a principal components analysis}

\item{scale}{A logic, default value is FALSE, but if mat is not be normalized, it should be TRUE.}
}
\value{
A prcomp object
}
\description{
performs a principal components analysis
}
\examples{
\dontrun{
library(bladderbatch)
data(bladderdata)
dat <- bladderEset[1:1000,]
edata <- exprs(dat)
res <- pca(t(edata))
}
}
