% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{pcc}
\alias{pcc}
\title{Pearson Correlation Coefficient (PCC)}
\usage{
pcc(target, feature)
}
\description{
This custom function calculates the Pearson Correlation Coefficient (PCC) between
two vectors: `target` (true values) and `feature` (predicted values). It is commonly
used to evaluate the linear relationship between predicted and true values in regression.
}
\keyword{internal}
