\name{peppercorr}
\alias{peppercorr}
\docType{data}

\title{ Correlations Between Pepper Variables}

\description{
The data give the correlations between 6 pepper variables. The data are taken from the
article published by Silva et al. (2013).
}

\usage{data(peppercorr)}

\format{
An object of class "matrix".
}

\source{
Silva et al. (2013) Path analysis in multicollinearity for fruit traits of pepper.
   \emph{Idesia}, 31:55-60.
}

\examples{
data(peppercorr)
print(peppercorr)

# End (not run)
}

\keyword{datasets}
