% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharse_gtf.R
\name{pharse_gtf}
\alias{pharse_gtf}
\title{pharse genes from local gtf file}
\usage{
pharse_gtf(genecode_gtf, genecode_entrezid)
}
\arguments{
\item{genecode_gtf}{character, file path to genecode GTF file}

\item{genecode_entrezid}{character, file path to genecode entrezid file}
}
\value{
a list, contains tx2gene, genes
}
\description{
prase txdb, gene genes from local gtf file
}
