% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_chtmap}
\alias{plot_chtmap}
\title{Plot Complex Heatmap with Class Splits}
\usage{
plot_chtmap(df_, table_class_, geneset_)
}
\arguments{
\item{df_}{A data frame containing expression data.}

\item{table_class_}{A named vector of class annotations.}

\item{geneset_}{A named list of gene sets.}
}
\value{
A Heatmap object from the ComplexHeatmap package.
}
\description{
Uses the ComplexHeatmap package to generate a detailed heatmap with class splits.
}
