% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrich_gsea.R
\name{plot_enrich_gsea}
\alias{plot_enrich_gsea}
\title{plot_enrich_gsea}
\usage{
plot_enrich_gsea(
  enrichment,
  show_category = 10,
  plot_type = "dot",
  filter_by = "-log10(p.adjust)",
  color_by = "NES",
  color_range = c("blue2", "red2"),
  size_by = "Count",
  size_range = c(2, 6),
  text_len_limit = 40,
  facet_by = NULL,
  facet_scales = "free_y"
)
}
\arguments{
\item{enrichment}{data.frame, from GSEA enrichment,
result can be merged from multi enrichment, grouping info can be viewed by facet}

\item{show_category}{integer, show enriched category number}

\item{plot_type}{character, one of dot, bar}

\item{filter_by}{character, filter show_category by which column,
one of -log10(p.adjust), NES, Count}

\item{color_by}{character, map point color, one of -log10(p.adjust), NES, Count}

\item{color_range}{two element character vector, mapping color range, from low to high,
for gsea plot, the value of NSE = 0 will be white color}

\item{size_by}{character, map point size, currently, only Count avaible for gseaResult}

\item{size_range}{two element numeric vector, mapping point size, from low to high
only works if plot_type is dot}

\item{text_len_limit}{integer, wrap y-axis text length}

\item{facet_by}{character, facet plot by which column in enrichment,
default avaible value is direction, user use add aditional column in enrichment}

\item{facet_scales}{character, facet scales, one of free_y, free_x, "free, fixed,
only works if facet_by is not NULL}
}
\value{
ggplot2 object
}
\description{
dotplot for enrich result from enrich_GSEA
}
