% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrich_ora.R
\name{plot_enrich_ora}
\alias{plot_enrich_ora}
\title{plot_enrich_ora}
\usage{
plot_enrich_ora(
  enrichment,
  show_category = 10,
  plot_type = "dot",
  axis_x = "GeneRatio",
  order_by = "-log10(p.adjust)",
  color_by = "-log10(p.adjust)",
  color_range = c("blue2", "red2"),
  size_by = "Count",
  size_range = c(2, 6),
  text_len_limit = 40,
  facet_by = NULL,
  facet_scales = "free_y"
)
}
\arguments{
\item{enrichment}{data.frame, from ORA enrichment,
result can be merged from multi enrichment, grouping info can be viewed by facet}

\item{show_category}{integer, show enriched category number}

\item{plot_type}{character, one of dot, bar}

\item{axis_x}{character, plot axis x by  which column,
one of GeneRatio, -log10(p.adjust), enrichFactor, Count}

\item{order_by}{character, filter show_category by which column,
one of GeneRatio, -log10(p.adjust), enrichFactor, Count}

\item{color_by}{character, map point color, one of GeneRatio, -log10(p.adjust), enrichFactor, Count}

\item{color_range}{two element character vector, mapping color range, from low to high}

\item{size_by}{character, map point size, one of GeneRatio, -log10(p.adjust), enrichFactor, Count
only works if plot_type is dot}

\item{size_range}{two element numeric vector, mapping point size, from low to high
only works if plot_type is dot}

\item{text_len_limit}{integer, wrap y-axis text length}

\item{facet_by}{character, facet plot by which column in enrichment}

\item{facet_scales}{character, facet scales, one of free_y, free_x, free, fixed,
only works if facet_by is not NULL}
}
\value{
ggplot2 object
}
\description{
dotplot for enrich result from enrich_ORA
}
