% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_GO}
\alias{plot_GO}
\title{Plot GO Terms}
\usage{
plot_GO(df_, title_, nTerm_ = 5, colorBy_ = "ONTOLOGY")
}
\arguments{
\item{df_}{A dataframe, result attr of a GO obj.}

\item{title_}{A character string specifying the title of the plot.}

\item{nTerm_}{An integer specifying the number of top terms to display (default is 5).}

\item{colorBy_}{A character string indicating which feature to group and color the bars by. Defaults to 'ONTOLOGY'.}
}
\value{
A `ggplot` object displaying the bar plot for the top GO terms, grouped and colored by the specified feature.
}
\description{
This function creates a bar plot for the top GO terms based on their -log10(q-value).
It supports grouping the data by a specified feature, such as the ontology type.
}
\examples{
plot_GO(res_GO@result, 'Test')

}
