% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_GOKEGG}
\alias{plot_GOKEGG}
\title{Plot Gene Ontology (GO) and KEGG Pathway Terms Together}
\usage{
plot_GOKEGG(go_, kegg_, title_, nTerm_ = 5, avoid_ = "")
}
\arguments{
\item{go_}{A GO result object containing the GO data.}

\item{kegg_}{A KEGG result object containing the KEGG pathway data.}

\item{title_}{A character string specifying the title of the plot.}

\item{nTerm_}{An integer specifying the number of top terms to display (default is 5).}

\item{avoid_}{A character string to exclude specific terms from the plot. Default is an empty string, meaning no terms are excluded.}
}
\value{
A `ggplot` object displaying the combined bar plot for the top GO and KEGG terms.
}
\description{
This function combines the GO and KEGG results and creates a bar plot for the top terms based on their -log10(q-value).
}
\examples{
plot_GOKEGG(res_GO, res_KEGG, 'Test')

}
