% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot Gene Set Enrichment Analysis (GSEA) Results}
\usage{
plot_gsea(gsea_, term_, sizef_ = 2, color_ = get_color(2, "ddlc"))
}
\arguments{
\item{gsea_}{A GSEA result object from the `clusterProfiler` package.}

\item{term_}{The term to visualize.}

\item{sizef_}{A scaling factor for the line thickness. Default is 2.}

\item{color_}{A vector of colors for the gradient scale.}
}
\value{
A ggplot object representing the GSEA plot.
}
\description{
Generates a GSEA plot for a specific term.
}
