% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot heatmap}
\usage{
plot_heatmap(
  matrix,
  group = NULL,
  scale_method = "scale",
  row_cluster_type = "auto",
  col_cluster_type = "auto",
  split_row_num = 2,
  color_map = c(-2, 0, 2),
  color_palette = c("#377eb8", "white", "#e41a1c"),
  row_names_size = 2,
  column_names_size = 8,
  dist_method = "euclidean",
  hclust_method = "average",
  ...
)
}
\arguments{
\item{matrix}{matrix, with unique rownames and colnames}

\item{group}{vector, used to anno column or cluster_withon_group, the order is same with matrix colname}

\item{scale_method}{character, scale method, one of "scale", "log1p", "none"}

\item{row_cluster_type}{character, cluster row method, one of "auto", "none"}

\item{col_cluster_type}{character, cluster column method, one of "auto", "semi", "none"}

\item{split_row_num}{integer, divide row cluster into how many sub-clusters, using cutree method}

\item{color_map}{numeric vector, heatmap color mapping cut, length must be same with color_palette}

\item{color_palette}{character vector, color mapped to color_map, length must be same with color_map}

\item{row_names_size}{integer, rownames size}

\item{column_names_size}{nteger, colnames size}

\item{dist_method}{dist method, refer to dist}

\item{hclust_method}{hclust method, refer to hclust}

\item{...}{additional parameters pass to ComplexHeatmap::Heatmap}
}
\value{
a list, contain heatmap and scaled ordered heatmap data
}
\description{
plot heatmap using complexHeatmap
}
