% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kegg}
\alias{plot_kegg}
\title{Visualize KEGG Pathway with Differential Gene Expression Data}
\usage{
plot_kegg(
  deg_,
  id_,
  logFC_ = "log2FoldChange",
  padj_ = "padj",
  color_ = color_myth
)
}
\arguments{
\item{deg_}{A data frame with gene expression results (e.g., log fold change and adjusted p-values).}

\item{id_}{The KEGG pathway ID (e.g., "hsa00010").}

\item{logFC_}{Column name for log fold change in `deg_`. Default is "log2FoldChange".}

\item{padj_}{Column name for adjusted p-values in `deg_`. Default is "padj".}

\item{color_}{A vector of colors for the gradient scale. Default is `color_myth`.}
}
\value{
A ggraph object visualizing the KEGG pathway.
}
\description{
This function overlays gene expression data onto a KEGG pathway graph.
}
