% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sc_deg}
\alias{plot_sc_deg}
\title{Plot Single-Cell Differential Expression (DEG) Results with GO Enrichment}
\usage{
plot_sc_deg(
  seu_,
  marker_all_,
  num_each_ = 15,
  ident_ = Idents(seu_),
  breaks_ = c(-5, 0, 5),
  organism_ = "hsa"
)
}
\arguments{
\item{seu_}{A Seurat object containing the single-cell RNA-seq data.}

\item{marker_all_}{A data frame with all marker gene data containing columns `gene`, `cluster`, and `avg_log2FC`.}

\item{num_each_}{An integer specifying the number of genes to display per cluster (default is 15).}

\item{ident_}{A vector of cluster identities to use in the analysis (default is the cluster identities from the Seurat object).}

\item{breaks_}{A numeric vector specifying the breaks for the color scale in the heatmap (default is c(-5, 0, 5)).}

\item{organism_}{A character string specifying the organism for GO enrichment analysis. Options are `'hsa'` for human (default) or `'mmu'` for mouse.}
}
\value{
A combined plot consisting of:
  - A heatmap showing the average log2 fold change for genes across clusters.
  - A dot plot showing the cluster annotations.
  - A bar plot showing the top 5 GO terms enriched in each cluster.
}
\description{
This function creates a combined visualization of single-cell differential expression (DEG) results.
It includes a heatmap of log2 fold changes, a dot plot of cluster annotations, and a bar plot for Gene Ontology (GO) enrichment results.
}
\examples{
plot_sc_deg(seurat_object, marker_data, num_each_ = 20, ident_ = c('Cluster1', 'Cluster2'), organism_ = 'hsa')

}
