% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.coloredChr}
\alias{print.coloredChr}
\title{Print Colored Character Vector with ANSI Colors}
\usage{
\method{print}{coloredChr}(x, ...)
}
\arguments{
\item{x}{An object of class `coloredChr`, which is a character vector containing color information.}

\item{...}{Additional arguments passed to the method (currently not used).}
}
\value{
No return value. The function prints a plot and the color vector in the console.
}
\description{
This function prints a character vector with ANSI colors in a plot and also displays the vector values in the console.
It is designed to work with objects of class `coloredChr`.
}
\examples{
vec = asColoredChr(c("red", "green", "blue"))
print(vec)  # This will display a plot and print the vector to the console.

}
