% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{produce_metric_regr}
\alias{produce_metric_regr}
\title{Produce Regression Metric Function}
\usage{
produce_metric_regr(metric_)
}
\arguments{
\item{metric_}{A string containing the name of the metric to be used for evaluation in mlr3.}
}
\value{
A function that computes the score for a given target and feature based on the metric.
}
\description{
This function produces a metric evaluation function based on the supplied
metric name, which can then be used to evaluate regression models.
}
\keyword{internal}
