% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDMS.R
\name{readDMS}
\alias{readDMS}
\title{Read DMS Raw Data}
\usage{
readDMS(path, name)
}
\arguments{
\item{path}{Character, single path to the tsv file containing raw DMS data.}

\item{name}{Character, name of the gene.}
}
\value{
Data frame containing the DMS data with the following headers: wt_aa pos mut_aa wt_codon mut_codon annotation   nonselect1   nonselect2
}
\description{
Read raw data from DMS experiment from a tsv file and format into a data frame.
}
\details{
A more detailed description of what the function does.
}
\examples{
# this example only works on my computer
setwd("/Users/Rachel/desktop/Roth Lab/R_DMS")
path <- './data/rawData_UBE2I_solid.txt'
gene_name <- 'UBE2I'
readDMS(path, gene_name)

}
\seealso{
\code{\link{another_function}} Info
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
