% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{readSeries}
\alias{readSeries}
\title{Read and Process a GEO Series Matrix File}
\usage{
readSeries(dir_, alwaysGZ_ = T)
}
\arguments{
\item{dir_}{A string specifying the directory containing the GEO Series Matrix file. The directory must contain exactly one file.}

\item{alwaysGZ_}{A boolean value indicating whether to always decompress gz-compressed txt files in the folder.}
}
\value{
A list with two components:
  - `expr`: A tidy tibble of expression data, where rows are samples, and columns are gene symbols.
  - `phen`: A tidy tibble of phenotype data, including metadata for each sample.
}
\description{
This function reads a GEO Series Matrix file from a specified directory, processes the expression
and phenotype data, and returns them in tidy formats.
}
\details{
- The function checks for and processes the platform annotation (GPL). If the annotation cannot be found,
  a warning is issued, and the gene IDs are retained as they appear in the expression data.
- Genes with the identifier `'permuted_negative'` are removed.
}
