% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAfold.R
\name{RNAfold}
\alias{RNAfold}
\title{Invoke RNAfold}
\usage{
RNAfold(RNA, options = "--noPS -p --jobs='0'")
}
\arguments{
\item{RNA}{Character vector of RNA sequences to be analyzed.}

\item{options}{Additional parameters to pass to RNAfold. Default is "--noPS -p --jobs='0'". (Calculates partition function and MFE structure). Warning: always use --noPS!}
}
\value{
The output of RNAfold as a character string.
}
\description{
Input RNA sequences to be analyzed by RNAfold.
}
\details{
Pass a character vector of RNA sequences to be analyzed (and optional additional parameters for RNA fold). Returns the output of RNA fold in string format.
Make sure to always keep the option --noPS. This funciton will also create dot.ps in the project directory (a dot plot). It will only create a single file if multiple inputs are supplied and continuously overwrite it. I don't want it to write this file but I can't seem to figure out how to turn it off, so the funcion jsut deletes this file before returning.
}
\examples{
RNAfold(c("AAAAUAUUUUUUUAUAUAUUUUUUUAAAAAAA", "CCCCCTTTTTCTATTTTAAAAAAAAUUUUUU"))


}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
