% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-summary.R
\name{simple_summary}
\alias{simple_summary}
\alias{simple_summary.default}
\alias{simple_summary.summary.coxph}
\title{simple_summary}
\usage{
simple_summary(summary_obj)

\method{simple_summary}{default}(summary_obj)

\method{simple_summary}{summary.coxph}(summary_obj)
}
\arguments{
\item{summary_obj}{The summary class}
}
\value{
a data.frame
}
\description{
This function would extract tidy information
}
