% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data.R
\name{split_expr}
\alias{split_expr}
\title{Clean the expression data from TCGA}
\usage{
split_expr(expr)
}
\arguments{
\item{expr}{data.frame, the raw expression matrix with colname of molecular id and individual ids.}
}
\value{
data.frame, the clean expression matrix with colname of molecular id and individual ids.
}
\description{
Delete the duplicated individual and rename the colnames.
}
\details{
nothing
}
\examples{
data(expr);expr = split_expr(expr)
}
