% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{symbolIn}
\alias{symbolIn}
\title{Retrieve Genes Associated with a GO Term}
\usage{
symbolIn(name_, GO_)
}
\arguments{
\item{name_}{A string specifying the name of the GO term.}

\item{GO_}{A list of GO data, typically obtained from `getGO()`.}
}
\value{
A vector of gene identifiers associated with the GO term.
}
\description{
This function retrieves the genes associated with a specified GO term.
}
\examples{
GO = getGO()
genes = symbolIn("cell cycle", GO_ = GO)

}
