% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{tab2pre_regr}
\alias{tab2pre_regr}
\title{Transform Target and Feature into Prediction Table for Regression}
\usage{
tab2pre_regr(target_, feature_)
}
\arguments{
\item{target_}{A vector of true values.}

\item{feature_}{A vector of predicted values.}
}
\value{
A `PredictionRegr` object containing the target and feature.
}
\description{
This function takes a target (true values) and a feature (predicted values) and
converts them into a `data.table` object to be used for regression scoring.
}
\keyword{internal}
