% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidyRowSums}
\alias{tidyRowSums}
\title{Calculate Row Sums and Add as a New Column}
\usage{
tidyRowSums(df_, name_, con_ = where(is.numeric))
}
\arguments{
\item{df_}{A data frame containing the data.}

\item{name_}{A string specifying the name of the new column where the row sums will be stored.}

\item{con_}{An expr about how to select columns.
Defaults to `where(is.numeric)`, which selects all numeric columns.}
}
\value{
A modified data frame with the new column containing row sums.
}
\description{
This function calculates the row sums of selected numeric columns in a data frame
and adds the result as a new column with a specified name.
}
\examples{
library(dplyr)
df = tibble(a = 1:3, b = 4:6, c = letters[1:3])
df |> tidyRowSums("row_sum")

}
