% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidyT}
\alias{tidyT}
\title{Transpose a tibble}
\usage{
tidyT(df_, id_new_, id_old_ = colnames(df_)[[1]])
}
\arguments{
\item{df_}{A data frame to be transposed.}

\item{id_new_}{A string specifying the name of the new column that will contain the original column names (now row names).}

\item{id_old_}{A string specifying the name of the column to be used as rownames before transposing.
Defaults to the first column of the input data frame.}
}
\value{
A tibble with the transposed data, including a column with the original column names.
}
\description{
This function transposes a data frame, converting its columns into rows,
and optionally renames the rownames column in the output tibble.
}
\examples{
df = tibble(id = c("row1", "row2"), a = c(1, 2), b = c(3, 4))
df |> tidyT("new_id")

}
