% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{withCal}
\alias{withCal}
\title{Evaluate an Expression within a Specific Environment}
\usage{
withCal(x_, expr_)
}
\arguments{
\item{x_}{An environment or data object in which the expression should be evaluated.}

\item{expr_}{An R expression to be evaluated within the context of `x_`.}
}
\value{
The result of evaluating the expression `expr_` within the environment `x_`.
}
\description{
This function evaluates an expression in the environment of a specified data object, which allows for customized scoping and evaluation.
}
\examples{
data.frame(a = 1:3, b = 4:6) |> withCal(a + b)

}
