% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEO.R
\name{writeSeries}
\alias{writeSeries}
\title{Write GEO Series Data to Disk}
\usage{
writeSeries(series_, dir_, RData = T, gse_ = NULL)
}
\arguments{
\item{series_}{A list containing two components:
- `expr`: A tibble of expression data.
- `phen`: A tibble of phenotype data.}

\item{dir_}{A string specifying the directory where the files will be written. If the directory does not exist, it will be created.}

\item{RData}{A logical value. If `TRUE`, the function saves the data as an RData file in addition to CSV files. Defaults to `TRUE`.}

\item{gse_}{(Optional) A string specifying a GEO Series accession ID (e.g., "GSE12345"). If provided, it is appended to the output filenames.}
}
\value{
None. The function writes files to the specified directory.
}
\description{
This function writes processed GEO Series data (expression and phenotype) to specified files in a given directory.
}
\details{
- The function creates the output directory if it does not exist.
- Outputs include:
  - `expr.csv`: Expression data in CSV format.
  - `phen.csv`: Phenotype data in CSV format.
  - `all.RData`: An RData file containing the entire dataset (optional).
}
