% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{major2minor}
\alias{major2minor}
\title{Update cell type assignments}
\usage{
major2minor(..., cl2ma_, attention_, remain2_ = NULL)
}
\arguments{
\item{...}{Expressions that map cluster numbers to target groups.}

\item{cl2ma_}{Named vector. Current mapping of clusters to cell types.}

\item{attention_}{Character. The target cell type to focus on.}

\item{remain2_}{Character. The target group for unassigned clusters. Defaults to be NULL.}
}
\value{
An updated named vector of cell type assignments.
}
\description{
Updates the mapping of clusters to cell types based on the given focus cell type.
Unspecified clusters are added to the target group.
}
\examples{
cl2ma = cluster2major(cluster_A = c(0, 1), cluster_B = c(2, 3), cluster_C = NULL, nCluster_ = 5)
major2minor(cluster_A1 = 0, cluster_A2 = 1, cl2ma_ = cl2ma, attention_ = "cluster_A")

}
