% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_KEGG}
\alias{plot_KEGG}
\title{Plot KEGG Terms}
\usage{
plot_KEGG(df_, title_, nTerm_ = 5)
}
\arguments{
\item{df_}{A dataframe, result attr of a KEGG obj.}

\item{title_}{A character string specifying the title of the plot.}

\item{nTerm_}{An integer specifying the number of top terms to display (default is 5).}
}
\value{
A `ggplot` object displaying the bar plot for the top KEGG pathway terms.
}
\description{
This function creates a bar plot for the top KEGG pathway terms
based on their -log10(q-value) for a given KEGG result object.
}
\examples{
plot_KEGG(res_KEGG@result, 'Test')

}
