% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_PCA}
\alias{plot_PCA}
\title{plot PCA}
\usage{
plot_PCA(mat, phe, group = "condition", ntop = 500, scale = FALSE)
}
\arguments{
\item{mat}{A data.frame, the expression matrix in which column is multiple variables and row is individual.}

\item{phe}{A data.frame, the phenotype information.}

\item{group}{A character, one colname of phe variable.}

\item{ntop}{A numeric, the top number of gene to performs a principal components analysis}

\item{scale}{A logic, default value is FALSE, but if mat is not be normalized, it should be TRUE.}
}
\value{
A ggplot object
}
\description{
plot PCA
}
\examples{
\dontrun{
library(bladderbatch)
data(bladderdata)
dat <- bladderEset[1:1000,]
pheno <- pData(dat)
edata <- exprs(dat)
plot_PCA(t(edata),phe=pheno,group="cancer")
}
}
