% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP visualization}
\usage{
plot_umap(df_, x_, y_, colorBy_)
}
\arguments{
\item{df_}{Data frame. Contains UMAP coordinates and grouping information.}

\item{x_}{Character. The column name for the UMAP x-axis.}

\item{y_}{Character. The column name for the UMAP y-axis.}

\item{colorBy_}{Character. The column name for the grouping variable used for coloring points.}
}
\value{
A ggplot object displaying the UMAP visualization.
}
\description{
Creates a UMAP plot with points colored by the specified grouping variable.
}
\examples{
plot_umap(df_, x_ = "UMAP_1", y_ = "UMAP_2", colorBy_ = "CellType")

}
