% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot volcano}
\usage{
plot_volcano(
  res,
  fc_thred = 2,
  padj_thred = 0.05,
  size = 1.5,
  alpha = 0.7,
  title = "Volcano of DEGs",
  xlim = NULL,
  ylim = NULL,
  ns_resampling = 1000,
  color = c("#0571b0", "#bababa", "#ca0020")
)
}
\arguments{
\item{res}{a data.drame from DESeq2 result}

\item{fc_thred}{foldchange cutoff, corelated to log2FoldChange column, high value mean less filter}

\item{padj_thred}{p.adjust cutoff, corelated to padj column, high value mean less filter}

\item{size}{point size}

\item{alpha}{point transparency}

\item{title}{plot title}

\item{xlim}{two element numeric vector, restrict x axis, default: NULL}

\item{ylim}{two element numeric vector, restrict y axis, default: NULL}

\item{ns_resampling}{numeric, downsampling NS points}

\item{color}{three element character vector, map point color, ordered by down-regulated, ns, up-regulated}
}
\value{
ggplot2 object
}
\description{
plot volcano plot for DESeq2 pariwise compare result
}
