% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Create a Volcano Plot}
\usage{
plot_volcano(
  df_,
  x_ = "logFC",
  y_ = "log10padj",
  label_ = "label",
  colorBy_ = "direction",
  colors_ = color_myth,
  alpha_ = 0.75,
  range_ = c(-12, 12)
)
}
\arguments{
\item{df_}{A data frame containing `x_`, `y_`, `label_`, and a grouping column (e.g., `colorBy_`).}

\item{x_}{Column name for the x-axis variable. Default is "logFC".}

\item{y_}{Column name for the y-axis variable. Default is "log10padj".}

\item{label_}{Column name for gene labels. Default is "label".}

\item{colorBy_}{Column name used for splitting data into "up", "down", and "non" groups. Default is "direction".}

\item{colors_}{Named vector of colors for "up", "down", and "non" categories.}

\item{alpha_}{Point transparency. Default is 0.75.}

\item{range_}{A numeric vector of length 2 specifying the x-axis range. Default is c(-12, 12).}
}
\value{
A ggplot object representing the volcano plot.
}
\description{
Combines up-, down-, and non-significant layers into a complete volcano plot.
}
\examples{
df = data.frame(
 logFC = c(-3, 1.5, 2, -4, 0),
 log10padj = c(1, 2, 3, 4, 0.5),
 label = c("Gene1", "Gene2", NA, "Gene4", NA),
 direction = c("down", "up", "up", "down", "non")
)
volcano_plot(df)

}
