% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showRunInfos}
\alias{showRunInfos}
\title{Show running information.}
\usage{
showRunInfos(msg, level = c("INFO", "WARN", "ERROR"), verbose = TRUE)
}
\arguments{
\item{level}{Specify the output level of the message, 'INFO', 'WARN' or 'ERROR', default: 'INFO'.}

\item{verbose}{Show running infos or not, default: TRUE.}

\item{Message}{needs to be shown.}
}
\description{
Show running information.
}
\examples{

logTransform(runif(n = 100, min = 1, max = 1000))
}
