% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFold.R
\name{varFold}
\alias{varFold}
\title{Function Template}
\usage{
varFold(wt_sequence, variant = FALSE, start = NULL, stop = NULL,
  wt = NULL, mut = NULL)
}
\arguments{
\item{variant}{The variant}

\item{wt_sequnece}{Character string representing the RNA sequence to calculate RNA structure for (DNA also ok... will substitute U for T)}
}
\value{
The sum of x and y
}
\description{
Description of the function
}
\details{
A more detailed description of what the function does.
}
\examples{

example_function(2,3)
# [1] 5




}
\seealso{
\code{\link{another_function}} Info
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
